feat(audio): add `audio` support module with `load_audio_mono_sr` implementation and integrate optional `hound` dependency
* Move `load_audio_mono_sr` function into a dedicated `audio` support module.
* Update `Cargo.toml` to include the new optional `audio` feature and `hound` dependency.
* Refactor tests in `silero_vad` to utilize the new `load_audio_mono_sr` module.
