+++ crates/bunsen/src/support/mod.rs
@@ -1,10 +1,12 @@
 #![allow(unused)]
 //! # Bunsen / Client Support Utilities
 
+pub mod arrays;
+pub mod math;
 pub mod validators;
 
 #[cfg(feature = "testing")]
 pub mod testing;
 
-pub mod arrays;
-pub mod math;
+#[cfg(feature = "audio")]
+pub mod audio;
